// src/app/core/models/role.ts
export enum Role {
  All = 'All',
  Admin = 'Admin',
  Doctor = 'Doctor',
  Specialite = 'Specialite',
  Genre = 'Genre',
  Service = 'Service',
  TypeEmploye = 'TypeEmploye',
  StatutDocument = 'StatutDocument',
  ArretMaladie = 'ArretMaladie',
  Reception = 'Reception',
  Comptable = 'Comptable',
  Assurance= 'Assurance',
  Infirmier='Infirmier',
  Medecin='Medecin',
  Pharmacien='Pharmacien',
  Informaticien='Informaticien',
  Logistique='Logistique'
}